update star ratings
